Рефакторинг в соотвитствии со стилем кода проекта
