+++ packages/client/src/App.tsx
@@ -1,11 +1,11 @@
 import { useAppDispatch, useAppSelector } from './store/hooks';
 import { useCallback, useEffect } from 'react';
-import { authActions, getAuthUserInfo } from './store/slices/auth';
+import { authActions, getAuthUserInfo } from './store/slices/auth/auth';
 import { AppRouter } from './providers/Router';
 import './App.css';
 
 function App() {
-  const inited = useAppSelector(state => state.auth._inited);
+  const isInited = useAppSelector(state => state.auth.isInited);
   const dispatch = useAppDispatch();
 
   const initFetch = useCallback(() => {
@@ -17,7 +17,7 @@ function App() {
     initFetch();
   }, []);
 
-  return <>{inited && <AppRouter />}</>;
+  return <>{isInited && <AppRouter />}</>;
 }
 
 export default App;
