+++ packages/client/src/pages/Profile/Avatar/Avatar.tsx
@@ -1,11 +1,13 @@
-import { useState } from 'react';
+import { FC, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { IAvatarProps } from './types';
 import { UserApi } from '../../../api/UserApi/UserApi';
+import { BASE_URL } from '../../../api/constants';
 
-export const Avatar = ({ avatar, fetchUserInfo }: IAvatarProps) => {
-  const [visibleFormAvatar, setVisibleFormAvatar] = useState(false);
+export const Avatar: FC<IAvatarProps> = ({ avatar, fetchUserInfo }) => {
+  const [isVisibleFormAvatar, setIsVisibleFormAvatar] = useState(false);
   const { register, handleSubmit } = useForm<{ picture: FileList }>();
+  const AVATAR_URL = `${BASE_URL}/resources${avatar}`;
 
   const onSubmit = (data: { picture: FileList }) => {
     const formData = new FormData();
@@ -23,20 +25,20 @@ export const Avatar = ({ avatar, fetchUserInfo }: IAvatarProps) => {
   };
 
   const handleVisibleFormAvatar = () => {
-    setVisibleFormAvatar(prev => !prev);
+    setIsVisibleFormAvatar(prev => !prev);
   };
 
   return (
     <div className="flex flex-col justify-center items-center flex-wrap">
       <div className="flex justify-center align-middle flex-wrap w-32 h-32 bg-slate-50 rounded-full relative overflow-hidden">
         <img
-          src={`https://ya-praktikum.tech/api/v2/resources${avatar}`}
+          src={AVATAR_URL}
           alt="avatar"
           className="max-h-full max-w-full hover:bg-opacity-1 rounded-full hover:opacity-70 cursor-pointer"
           onClick={handleVisibleFormAvatar}
         />
       </div>
-      {visibleFormAvatar && (
+      {isVisibleFormAvatar && (
         <form onSubmit={handleSubmit(onSubmit)}>
           <input {...register('picture')} type="file" />
           <button className="text-white btn-primary">Отправить</button>
