+++ packages/client/src/pages/Profile/Avatar/types.ts
@@ -1,4 +1,4 @@
 export interface IAvatarProps {
   avatar: string | null;
-  fetchUserInfo: () => void;
+  fetchUserInfo: VoidFunction;
 }
